fix: move user pref sections under ../settings/userPreferences/*
