+++ apps/desktop/src/lib/settings/userPreferences/DetailsForm.svelte
